refactor: Extract Leave Request Item component
- make ion-modal style global
