+++ frontend/src/main.css
@@ -1,2 +1,6 @@
 @import "./assets/Inter/inter.css";
 @import "frappe-ui/src/style.css";
+
+ion-modal {
+	--height: auto;
+}
